inference_server/pyproject.toml
[tool.poetry]
name = "inference_server"
version = "0.1.0"
description = ""
authors = ["sovitpy <sovitlekhak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
langchain-groq = "^0.1.9"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
requests = "^2.32.3"
qdrant-client = "^1.11.3"
sentence-transformers = "^3.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
